+++ encoding/thrift/doc.go
@@ -32,14 +32,9 @@
 //
 // 	thriftrw --plugin yarpc myservice.thrift
 //
-// Pass "-gomock" to the plugin to generate gomock-compatible mock clients for
-// the Thrift services.
-//
-// 	thriftrw --plugin "yarpc -gomock" myservice.thrift
-//
 // In addition to generating code for types specified in your Thrift file,
 // this will generate two packages for each service in the file: a client
-// package, a server package, and if you requested mocks, a test package.
+// package, a server package, and a test package.
 //
 // 	myservice
 // 	   |- myserviceclient
@@ -56,7 +51,7 @@
 // 	handler := myHandler{}
 // 	dispatcher.Register(myserviceserver.New(handler))
 //
-// The test package provides a gomock-compatible mock client.
+// The test package provides a gomock-compatible mock client for the service.
 //
 // 	mockCtrl := gomock.NewController(t)
 // 	client := myservicetest.NewMockClient(mockCtrl)